+++ submit-api/src/submit_api/resources/activity_log.py
@@ -18,9 +18,11 @@ from http import HTTPStatus
 from flask_cors import cross_origin
 from flask_restx import Namespace, Resource
 
+from submit_api.auth import jwt
 from submit_api.resources.apihelper import Api as ApiHelper
 from submit_api.schemas.activity_log import ActivityLogSchema
 from submit_api.services.activity_log_service import ActivityLogService
+from submit_api.utils.roles import EpicSubmitRole
 from submit_api.utils.util import allowedorigins, cors_preflight
 
 
@@ -50,6 +52,7 @@ class ActivityLog(Resource):
     def get(entity_type, entity_id):
         """Retrieve activity logs for a specific entity type and ID."""
         # Retrieve logs
+        is_staff = jwt.contains_role([EpicSubmitRole.EAO_VIEW.value])
         logs = ActivityLogService.get_activity_logs(entity_type, entity_id)
-        schema = ActivityLogSchema(many=True, context={"is_proponent": True})
+        schema = ActivityLogSchema(many=True, context={"is_proponent": not is_staff})
         return schema.dump(logs), HTTPStatus.OK
