+++ submit-api/src/submit_api/resources/invitation.py
@@ -23,11 +23,12 @@ from submit_api.auth import auth
 from submit_api.enums.role import ProponentPermissionsEnum
 from submit_api.resources.apihelper import Api as ApiHelper
 from submit_api.schemas.account import AccountCreateSchema
-from submit_api.schemas.invitation import (InvitationSchema, CreateInvitationToExistingAccountProjectSchema,
-                                           CreateNewAccountInvitationSchema)
+from submit_api.schemas.invitation import (
+    CreateInvitationToExistingAccountProjectSchema, CreateNewAccountInvitationSchema, InvitationSchema)
 from submit_api.services.invitation_service import InvitationService
 from submit_api.utils.roles import EpicSubmitRole
 from submit_api.utils.util import allowedorigins, cors_preflight
+
 
 API = Namespace("invitations", description="Endpoints for Invitation Management")
