+++ submit-api/src/submit_api/resources/staff_user.py
@@ -22,10 +22,11 @@ from flask_restx import Namespace, Resource
 from submit_api.auth import auth
 from submit_api.exceptions import ResourceNotFoundError
 from submit_api.resources.apihelper import Api as ApiHelper
-from submit_api.schemas.staff_user import StaffUserSchema, CreateStaffUserRequest
+from submit_api.schemas.staff_user import CreateStaffUserRequest, StaffUserSchema
 from submit_api.services.staff_user_service import StaffUserService
 from submit_api.utils.roles import EpicSubmitRole
 from submit_api.utils.util import allowedorigins, cors_preflight
+
 
 API = Namespace("staff-user", description="Endpoints for Staff Management")
 """Custom exception messages
