+++ submit-api/src/submit_api/resources/submission.py
@@ -23,7 +23,9 @@ from submit_api.auth import auth
 from submit_api.resources.apihelper import Api as ApiHelper
 from submit_api.schemas.submission import CreateSubmissionRequestSchema, SubmissionSchema
 from submit_api.services.submission import SubmissionService
+from submit_api.utils.roles import EpicSubmitRole
 from submit_api.utils.util import allowedorigins, cors_preflight
+
 
 API = Namespace("submissions", description="Endpoints for Submission Management")
 """Custom exception messages
@@ -132,6 +134,25 @@ class DocumentSubmission(Resource):
     def delete(submission_id):
         """Delete a submission document."""
         deleted_submission = SubmissionService.delete_submission(submission_id)
+        return SubmissionSchema().dump(deleted_submission), HTTPStatus.OK
+
+
+@cors_preflight("OPTIONS, DELETE")
+@API.route("/<int:submission_id>/document/soft-delete", methods=["OPTIONS", "DELETE"])
+class DocumentSubmissionSoftDelete(Resource):
+    """Resource for managing a document submission."""
+
+    @staticmethod
+    @ApiHelper.swagger_decorators(API, endpoint_description="Soft delete a document submission")
+    @API.response(
+        code=HTTPStatus.OK, model=submission_model, description="Submission"
+    )
+    @API.response(HTTPStatus.BAD_REQUEST, "Bad Request")
+    @cross_origin(origins=allowedorigins())
+    @auth.has_one_of_staff_roles([EpicSubmitRole.EAO_EDIT.value])
+    def delete(submission_id):
+        """Soft delete a submission document."""
+        deleted_submission = SubmissionService.soft_delete_submission(submission_id)
         return SubmissionSchema().dump(deleted_submission), HTTPStatus.OK
